+++ client/src/index.css
@@ -18,7 +18,7 @@ body {
   margin: auto;
   padding: 0;
   box-sizing: border-box;
-  width: 60vw;
+  width: 100vw;
   min-height: 100vh;
 }
